pyproject.toml
[project]
name = "semantic-kernel-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "semantic-kernel>=1.32.2",
    "ioa_observe_sdk==1.0.12",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "a2a-sdk>=0.2.13",
    "python-a2a>=0.5.9",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0",
]
